refactor: componentsの配置を修正
