+++ src/components/copyright/index.tsx
@@ -1,11 +1,9 @@
 import styles from "./index.module.sass";
 
-const Copyright: React.FC = () => {
+export const Copyright: React.FC = () => {
   return (
     <div className={styles.inner}>
       <p className={styles.content}>©2021 rurito0125 All Rights Reserved.</p>
     </div>
   );
 };
-
-export default Copyright;
