+++ src/pages/index.tsx
@@ -3,8 +3,7 @@ import Profile from "./profile/index";
 import Skill from "./skill/index";
 import History from "./history/index";
 import Timeline from "./timeline/index";
-import Copyright from "../components/copyright/index";
-import { LoadingAnimation } from "../components/";
+import { LoadingAnimation, Copyright } from "../components/";
 
 const Home: React.FC = () => {
   return (
